+++ core/deployment/src/main/java/io/quarkus/deployment/dev/IDEDevModeMain.java
@@ -7,12 +7,14 @@ import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.BiConsumer;
+import java.util.stream.Collectors;
 
 import org.jboss.logging.Logger;
 
 import io.quarkus.bootstrap.BootstrapGradleException;
 import io.quarkus.bootstrap.app.CuratedApplication;
 import io.quarkus.bootstrap.model.AppArtifactKey;
+import io.quarkus.bootstrap.model.AppDependency;
 import io.quarkus.bootstrap.resolver.AppModelResolverException;
 import io.quarkus.bootstrap.resolver.QuarkusGradleModelFactory;
 import io.quarkus.bootstrap.resolver.maven.workspace.LocalProject;
@@ -34,13 +36,13 @@ public class IDEDevModeMain implements BiConsumer<CuratedApplication, Map<String
         try {
             if (BuildToolHelper.isMavenProject(appClasses)) {
                 LocalProject project = LocalProject.loadWorkspace(appClasses);
-                DevModeContext.ModuleInfo root = toModule(project);
+                DevModeContext.ModuleInfo root = toModule(project, curatedApplication);
                 devModeContext.setApplicationRoot(root);
                 for (Map.Entry<AppArtifactKey, LocalProject> module : project.getWorkspace().getProjects().entrySet()) {
                     if (module.getKey().equals(project.getKey())) {
                         continue;
                     }
-                    devModeContext.getAdditionalModules().add(toModule(module.getValue()));
+                    devModeContext.getAdditionalModules().add(toModule(module.getValue(), curatedApplication));
                 }
             } else {
                 // TODO find a way to reuse the previously model instead of building a new one.
@@ -48,11 +50,11 @@ public class IDEDevModeMain implements BiConsumer<CuratedApplication, Map<String
                         BuildToolHelper.getBuildFile(appClasses, BuildToolHelper.BuildTool.GRADLE).toFile(),
                         QuarkusModelHelper.DEVMODE_REQUIRED_TASKS);
                 final WorkspaceModule launchingModule = quarkusModel.getWorkspace().getMainModule();
-                DevModeContext.ModuleInfo root = toModule(quarkusModel.getWorkspace().getMainModule());
+                DevModeContext.ModuleInfo root = toModule(quarkusModel.getWorkspace().getMainModule(), curatedApplication);
                 devModeContext.setApplicationRoot(root);
                 for (WorkspaceModule additionalModule : quarkusModel.getWorkspace().getAllModules()) {
                     if (!additionalModule.getArtifactCoords().equals(launchingModule.getArtifactCoords())) {
-                        devModeContext.getAdditionalModules().add(toModule(additionalModule));
+                        devModeContext.getAdditionalModules().add(toModule(additionalModule, curatedApplication));
                     }
                 }
             }
@@ -65,7 +67,8 @@ public class IDEDevModeMain implements BiConsumer<CuratedApplication, Map<String
                 Collections.singletonMap(DevModeContext.class.getName(), devModeContext));
     }
 
-    private DevModeContext.ModuleInfo toModule(WorkspaceModule module) throws BootstrapGradleException {
+    private DevModeContext.ModuleInfo toModule(WorkspaceModule module, CuratedApplication application)
+            throws BootstrapGradleException {
         AppArtifactKey key = new AppArtifactKey(module.getArtifactCoords().getGroupId(),
                 module.getArtifactCoords().getArtifactId(), module.getArtifactCoords().getClassifier());
 
@@ -85,10 +88,12 @@ public class IDEDevModeMain implements BiConsumer<CuratedApplication, Map<String
                 module.getSourceSet().getResourceDirectory().getPath(),
                 sourceParents,
                 module.getBuildDir().toPath().resolve("generated-sources").toAbsolutePath().toString(),
-                module.getBuildDir().toString());
+                module.getBuildDir().toString(),
+                application.getAppModel().getUserDependencies().stream().map(AppDependency::getArtifact)
+                        .collect(Collectors.toList()));
     }
 
-    private DevModeContext.ModuleInfo toModule(LocalProject project) {
+    private DevModeContext.ModuleInfo toModule(LocalProject project, CuratedApplication application) {
         return new DevModeContext.ModuleInfo(project.getKey(), project.getArtifactId(),
                 project.getDir().toAbsolutePath().toString(),
                 Collections.singleton(project.getSourcesSourcesDir().toAbsolutePath().toString()),
@@ -96,6 +101,8 @@ public class IDEDevModeMain implements BiConsumer<CuratedApplication, Map<String
                 project.getResourcesSourcesDir().toAbsolutePath().toString(),
                 project.getSourcesDir().toString(),
                 project.getCodeGenOutputDir().toString(),
-                project.getOutputDir().toString());
+                project.getOutputDir().toString(),
+                application.getAppModel().getUserDependencies().stream().map(AppDependency::getArtifact)
+                        .collect(Collectors.toList())); //TODO: how can we resolve the per module class path?
     }
 }
