+++ core/deployment/src/main/java/io/quarkus/deployment/dev/IsolatedDevModeMain.java
@@ -193,9 +193,9 @@ public class IsolatedDevModeMain implements BiConsumer<CuratedApplication, Map<S
                 compilationProviders.add(provider);
                 context.getAllModules().forEach(moduleInfo -> moduleInfo.addSourcePaths(provider.handledSourcePaths()));
             }
-            ClassLoaderCompiler compiler;
+            DevModeCompiler compiler;
             try {
-                compiler = new ClassLoaderCompiler(Thread.currentThread().getContextClassLoader(), curatedApplication,
+                compiler = new DevModeCompiler(curatedApplication,
                         compilationProviders, context);
             } catch (Exception e) {
                 log.error("Failed to create compiler, runtime compilation will be unavailable", e);
